feat: add support for passing access tokens, or access token credential files to the keyring as environment variables
Sometimes, there are scenarios where keyring needs to be invoked from a docker container, and you would
probably not want to install 'gcloud' in your docker container to keep it light weight. in these
cases, its easier for the developer to create the access config file using gcloud outside the
docker build, and then passing them as a docker secret.

a specific scenario: while using google cloud build along with docker buildkit, the google auth keyring
does not work, because the docker builder needs to be building on --network=cloudbuild, but
--network=cloudbuild is an unsupported network on docker buildkit builder